+++ frontend/src/page/channelPost/DetailPost.js
@@ -7,6 +7,7 @@ import ChannelPostService from '../../api/service/ChannelPostService'
 import CommentFrame from '../../component/comment/CommentFrame'
 import {useNavigate} from "react-router-dom";
 import VoteService from "../../api/service/VoteService";
+import ChannelService from "../../api/service/ChannelService";
 
 const DetailPost = () => {
     const navigate = useNavigate();
@@ -59,11 +60,12 @@ const DetailPost = () => {
     }
 
     const handleDeleteButton = async () => {
-        let result = await ChannelPostService.delete(postId, channelId);
-
-        if (result) {
-            alert('삭제되었습니다.')
-            window.location.href = '/channel/' + channelId + '?page=1'
+        if (window.confirm('삭제하시겠습니까?')) {
+            let result = await ChannelPostService.delete(postId, channelId);
+            if (result) {
+                alert('삭제되었습니다.')
+                window.location.href = '/channel/' + channelId + '?page=1'
+            }
         }
     }
 
@@ -75,7 +77,7 @@ const DetailPost = () => {
     }
 
     const authorize = (post) => {
-        let memberId = sessionStorage.getItem('id')
+        let memberId = localStorage.getItem('id')
 
         if (post && Number(memberId) === post.memberId) {
             setAuthentication(true)
@@ -84,7 +86,7 @@ const DetailPost = () => {
 
     const createVote = async () => {
         let voteRequest = {
-            'memberId' : sessionStorage.getItem('id'),
+            'memberId' : localStorage.getItem('id'),
             'targetId' : post.id,
             'voteType' : 'ARTICLE',
         }
